[Multimedia] Moved enums of AudioManager into separate files.
