+++ src/Tizen.Multimedia/AudioManager/AudioStreamType.cs
@@ -1,0 +1,75 @@
+/*
+ * Copyright (c) 2016 Samsung Electronics Co., Ltd All Rights Reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the License);
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an AS IS BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Tizen.Multimedia
+{
+    /// <summary>
+    /// Specifies the audio stream types.
+    /// </summary>
+    /// <since_tizen> 3 </since_tizen>
+    public enum AudioStreamType
+    {
+        /// <summary>
+        /// Media.
+        /// </summary>
+        Media,
+
+        /// <summary>
+        /// System.
+        /// </summary>
+        System,
+
+        /// <summary>
+        /// Alarm.
+        /// </summary>
+        Alarm,
+
+        /// <summary>
+        /// Notification.
+        /// </summary>
+        Notification,
+
+        /// <summary>
+        /// Emergency.
+        /// </summary>
+        Emergency,
+
+        /// <summary>
+        /// Voice information.
+        /// </summary>
+        VoiceInformation,
+
+        /// <summary>
+        /// Voice recognition.
+        /// </summary>
+        VoiceRecognition,
+
+        /// <summary>
+        /// Ringtone for VoIP.
+        /// </summary>
+        RingtoneVoip,
+
+        /// <summary>
+        /// VoIP.
+        /// </summary>
+        Voip,
+
+        /// <summary>
+        /// Media only for external devices.
+        /// </summary>
+        MediaExternalOnly
+    }
+}
